move test script over to test_utils and continued to tweak function and fix test functions
